+++ store/toast/atom.ts
@@ -1,5 +1,5 @@
 import { atom } from 'recoil';
-import { ToastType } from '@/shared/type/global';
+import { ToastType } from '@/shared/type/common';
 
 export const toastStateAtom = atom<ToastType>({
   key: 'toastState',
